TAffineのLookAt系APIの仕様が変わりました。詳しくはAffine.hのコメントを見てください。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@3911 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Base/TMatrix.h
@@ -786,7 +786,7 @@ public:
 	void input(std::istream& is){
 		char ch;
 		is >> ch;
-		for(int i=0; i<height(); ++i) is >> row(i);
+		for(int i=0; i<(int)height(); ++i) is >> row(i);
 		is >> ch;
 	}
 	//@}
